[Android] Add message passing support
XWalkExtensionBridge.java is the interface to allow implement message
passing based extension written by java.

When new a instance that inherit from XWalkExtensionBridge.java, it will
create a C++ XWalkExtension and a XWalkExtension::Context, they will
bridge the messages passed between JavaScript and Java.

XWalkExtensionBridge is self-registed to XWalkExtensionService. To
achieve this, XWalkExtensionService member in XWalkBrowserMainParts is
defined as static member This is a short-term solution and only affect
Android port.

NEXT:
1. XWalkExtension lists logic decoupled from XWalkExtensionService.
   A new class XWalkExtensionList with following ability:
   * Singleton
   * GetExtensionList() return latest extension list
   * AddListner(XWalkExtensionService*) to allow send update when an
     extension get added and removed.
   * RegisterExtension() to add an extension and notify listeners
   * UnRegisterExtension() to remove an extension and notify listeners
2. HandleSyncMessage not working