Make XWalkModuleSystem easily access its v8::Context
Keep a v8::Persistent to get from the XWalkModuleSystem to its related
v8::Context. This won't cause a circular dependency because we are
called in WillReleaseScriptContext() when v8::Context will be released,
and at that point we destroying the Module System (disposing the
persistent handle).

This makes use of this in XWalkExtensionModules, that now take an
ModuleSystem and use it to get the right v8::Context when calling back
JS from C++, instead of expecting v8::Context to be passed.

This will be useful when we dissociate calling back JavaScript from
XWalkExtensionRenderViewHandler.